# Conflicts:
#	src/main/java/hu/oe/nik/szfmv/Main.java
#	src/main/java/hu/oe/nik/szfmv/automatedcar/Driver.java
#	src/main/java/hu/oe/nik/szfmv/automatedcar/bus/Signal.java
#	src/main/java/hu/oe/nik/szfmv/automatedcar/powertrainsystem/PowertrainSystem.java
#	src/main/java/hu/oe/nik/szfmv/environment/model/CollidableObject.java
#	src/main/java/hu/oe/nik/szfmv/environment/model/MovingObject.java
#	src/main/java/hu/oe/nik/szfmv/environment/model/WorldObject.java
#	src/main/java/hu/oe/nik/szfmv/environment/model/WorldObjectCollection.java
#	src/main/java/hu/oe/nik/szfmv/environment/object/Car.java
#	src/main/java/hu/oe/nik/szfmv/environment/object/Cyclist.java
#	src/main/java/hu/oe/nik/szfmv/environment/object/IDriveable.java
#	src/main/java/hu/oe/nik/szfmv/environment/object/Pedestrian.java
#	src/main/java/hu/oe/nik/szfmv/environment/object/Road.java
#	src/main/java/hu/oe/nik/szfmv/environment/object/RoadSign.java
#	src/main/java/hu/oe/nik/szfmv/environment/object/Tree.java
#	src/main/java/hu/oe/nik/szfmv/environment/util/ModelShape.java
#	src/main/java/hu/oe/nik/szfmv/visualisation/Coord.java
#	src/main/java/hu/oe/nik/szfmv/visualisation/CourseDisplay.java
#	src/main/java/hu/oe/nik/szfmv/visualisation/GameDisplayJPanel.java
#	src/main/java/hu/oe/nik/szfmv/visualisation/RoadConstants.java
#	src/main/java/hu/oe/nik/szfmv/visualisation/WorldObjectDisplayState.java

+++ src/main/java/hu/oe/nik/szfmv/visualisation/WorldObjectDisplayState.java
@@ -8,10 +8,10 @@ import java.awt.geom.AffineTransform;
 class WorldObjectDisplayState {
     private WorldObject wo;
     private Coord c;
-    private float rot;
+    private double rot;
     private AffineTransform t;
 
-    private WorldObjectDisplayState(WorldObject wo, Coord c, float rot, AffineTransform t) {
+    private WorldObjectDisplayState(WorldObject wo, Coord c, double rot, AffineTransform t) {
         this.wo = wo;
         this.c = c;
         this.rot = rot;
@@ -19,10 +19,7 @@ class WorldObjectDisplayState {
     }
 
     public static WorldObjectDisplayState createState(WorldObject object, AffineTransform t) {
-        return new WorldObjectDisplayState(
-                object,
-                new Coord(object),
-                object.getRotation(), t);
+        return new WorldObjectDisplayState(object, new Coord(object), object.getRotation(), t);
     }
 
     public AffineTransform getTransform() {
@@ -30,9 +27,7 @@ class WorldObjectDisplayState {
     }
 
     public boolean isChanged() {
-        return !(wo.getX() == c.getX()
-                && wo.getY() == c.getY()
-                && wo.getRotation() == rot);
+        return !(wo.getX() == c.getX() && wo.getY() == c.getY() && wo.getRotation() == rot);
     }
 
     public void updateState(AffineTransform t) {
